+++ routes/api/userRoutes.js
@@ -1,5 +1,5 @@
 const router = require("express").Router();
-const usersController = require("../../controllers/usersController");
+const usersController = require("../../controllers/userController");
 
 // Matches with "/api/users"
 router.route("/")
